+++ library/components/index.ts
@@ -20,6 +20,11 @@ import Menu from './menu/Menu.vue';
 import ButtonSearch from './buttonsearch/ButtonSearch.vue';
 import InputRangeNumber from './inputrangenumber/InputRangeNumber.vue';
 import Calendar from './calendar/Calendar.vue';
+import InputText from './inputtext/InputText.vue';
+import Form from './form/Form.vue';
+import DialogForm from './dialogform/DialogForm.vue';
+import InputNumber from './inputnumber/InputNumber.vue';
+import InputEmail from './inputemail/InputEmail.vue';
 
 export {
   Badge,
@@ -44,4 +49,9 @@ export {
   ButtonSearch,
   InputRangeNumber,
   Calendar,
+  InputText,
+  Form,
+  DialogForm,
+  InputNumber,
+  InputEmail,
 };