+++ playground/docs/dialog/DialogFormDocs.vue
@@ -4,6 +4,8 @@ import Card from 'lib/components/card/Card.vue';
 import DialogForm from 'lib/components/dialogform/DialogForm.vue';
 import { FormValue } from 'lib/components/form/Form.vue.d';
 import InputText from 'lib/components/inputtext/InputText.vue';
+import InputEmail from 'lib/components/inputemail/InputEmail.vue';
+import InputNumber from 'lib/components/inputnumber/InputNumber.vue';
 import { ref, shallowRef } from 'vue';
 import DocTitle from '../DocTitle.vue';
 
@@ -62,6 +64,18 @@ const apply = (e: {
             placeholder="Use custom placholder"
             use-validator
           />
+          <InputEmail
+            :validator-message="{
+              exceed: 'Max char 30',
+              empty: 'Should not empty',
+              invalidFormat: 'Formatnya salah',
+            }"
+            field-name="email"
+            label="Email"
+            mandatory
+            placeholder="Masukkan email"
+            use-validator
+          />
           <InputText
             :validator-message="{
               exceed: 'Max char 30',
@@ -79,6 +93,16 @@ const apply = (e: {
             }"
             field-name="lastname"
             label="Nama Belakang"
+            mandatory
+            use-validator
+          />
+          <InputNumber
+            :validator-message="{
+              exceed: 'Max char 30',
+              empty: 'Should not empty',
+            }"
+            field-name="phone"
+            label="Nomor Telepon"
             mandatory
             use-validator
           />